Refactor: Move `sendPlayerToServer` and `getServer` methods to `PluginMessageHelper` for better code organization.
